requirements.txt
flask
requests
python-dotenv
uv
